+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: sgury <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/13 09:28:29 by sgury             #+#    #+#             */
-/*   Updated: 2019/05/29 15:05:16 by flbeaumo         ###   ########.fr       */
+/*   Updated: 2019/05/29 19:25:23 by flbeaumo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ void			ft_usage(char conv);
 int			ft_dispatcher(va_list ap, t_data_tab *data, t_buff *buff);
 void			ft_check_flags(t_data_tab *data);
 char			*ft_get_str(va_list ap, t_data_tab *data);
-int			ft_get_nbr(va_list ap, t_data_tab *data);
+long long int		ft_get_nbr(va_list ap, t_data_tab *data);
 int			ft_get_unbr(va_list ap, t_data_tab *data);
 void			ft_size_error(t_data_tab *data);
